+++ include/metalang99/control.h
@@ -78,21 +78,22 @@
  * @code
  * #include <metalang99/control.h>
  *
- * #define X_IMPL(...)    M_overload(v(X_), v(__VA_ARGS__))
- * #define X_1_IMPL(a)    v(Billie & a)
- * #define X_2_IMPL(a, b) v(Jean & a & b)
+ * #define X(...)    M_overload(X_, __VA_ARGS__)
+ * #define X_1(a)    Billie & a
+ * #define X_2(a, b) Jean & a & b
  *
  * // Billie & 4
- * M_call(X, v(4))
+ * X(4)
  *
  * // Jean & 5 & 6
- * M_call(X, v(5, 6))
+ * X(5, 6)
  * @endcode
  *
- * @note This function calls @p f with #METALANG99_call, so no partial application occurs, and so
- * arity specifiers are not needed.
+ * @note @p f need not be postfixed with `_IMPL`. It is literally invoked as `<f><count of
+ * ...>(...)`.
  */
-#define METALANG99_overload(f, ...) METALANG99_call(METALANG99_overload, f, __VA_ARGS__)
+#define METALANG99_overload(f, ...)                                                                \
+    METALANG99_catPlain(f, METALANG99_PRIV_VARIADICS_COUNT(__VA_ARGS__))(__VA_ARGS__)
 
 /**
  * The plain version of #METALANG99_if.
@@ -111,48 +112,32 @@
  * The plain version of #METALANG99_whenLazy.
  */
 #define METALANG99_whenLazyPlain(cond, f) METALANG99_ifPlain(cond, f, METALANG99_consumePlain)
-
-/**
- * The plain version of #METALANG99_overload.
- *
- * @note @p f need not be postfixed with `_IMPL`. It is literally invoked as `<f><count of
- * ...>(...)`.
- */
-#define METALANG99_overloadPlain(f, ...)                                                           \
-    METALANG99_catPlain(f, METALANG99_PRIV_VARIADICS_COUNT(__VA_ARGS__))(__VA_ARGS__)
 // }
 
 #ifndef DOXYGEN_IGNORE
 
 // Implementation {
-#define METALANG99_overload_IMPL(f, ...)                                                           \
-    METALANG99_callTrivial(                                                                        \
-        METALANG99_catPlain(f, METALANG99_PRIV_VARIADICS_COUNT(__VA_ARGS__)),                      \
-        __VA_ARGS__)
-
-#define METALANG99_if_IMPL(cond, x, y) v(METALANG99_ifPlain(cond, x, y))
-
+#define METALANG99_if_IMPL(cond, x, y)    v(METALANG99_ifPlain(cond, x, y))
 #define METALANG99_when_IMPL(cond, x)     v(METALANG99_whenPlain(cond, x))
 #define METALANG99_whenLazy_IMPL(cond, f) v(METALANG99_ifPlain(cond, f, METALANG99_consume))
 // }
 
 // Arity specifiers {
-#define METALANG99_overload_ARITY 2
 #define METALANG99_if_ARITY       3
 #define METALANG99_when_ARITY     2
 #define METALANG99_whenLazy_ARITY 2
+#define METALANG99_overload_ARITY 2
 // }
 
 // Aliases {
 #ifndef METALANG99_NO_SMALL_PREFIX
 
-#define M_overload METALANG99_overload
 #define M_if       METALANG99_if
 #define M_when     METALANG99_when
 #define M_whenLazy METALANG99_whenLazy
+#define M_overload METALANG99_overload
 
 #define M_ifPlain       METALANG99_ifPlain
-#define M_overloadPlain METALANG99_overloadPlain
 #define M_whenPlain     METALANG99_whenPlain
 #define M_whenLazyPlain METALANG99_whenLazyPlain